refactor HandlerFactory
