+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/handlers/v2/HandlerConstructor.java
@@ -3,24 +3,25 @@ package com.sap.cloud.lm.sl.mta.handlers.v2;
 import static com.sap.cloud.lm.sl.common.util.CommonUtil.cast;
 
 import com.sap.cloud.lm.sl.mta.model.SystemParameters;
-import com.sap.cloud.lm.sl.mta.model.v1.DeploymentDescriptor;
-import com.sap.cloud.lm.sl.mta.model.v1.Platform;
+import com.sap.cloud.lm.sl.mta.model.v2.DeploymentDescriptor;
+import com.sap.cloud.lm.sl.mta.model.v2.Platform;
 import com.sap.cloud.lm.sl.mta.resolvers.PlaceholderResolver;
 import com.sap.cloud.lm.sl.mta.resolvers.Resolver;
 import com.sap.cloud.lm.sl.mta.resolvers.ResolverBuilder;
 import com.sap.cloud.lm.sl.mta.resolvers.v2.DescriptorPlaceholderResolver;
 import com.sap.cloud.lm.sl.mta.resolvers.v2.DescriptorReferenceResolver;
 
-public class HandlerConstructor extends com.sap.cloud.lm.sl.mta.handlers.v1.HandlerConstructor {
+public class HandlerConstructor implements com.sap.cloud.lm.sl.mta.handlers.HandlerConstructor{
+    
+    protected DescriptorHandler handler;
+
+    public HandlerConstructor() {
+        handler = getDescriptorHandler();
+    }
 
     @Override
     public DescriptorParser getDescriptorParser() {
         return new DescriptorParser();
-    }
-
-    @Override
-    public ConfigurationParser getConfigurationParser() {
-        return new ConfigurationParser();
     }
 
     @Override
@@ -29,6 +30,11 @@ public class HandlerConstructor extends com.sap.cloud.lm.sl.mta.handlers.v1.Hand
             return (DescriptorHandler) this.handler;
         }
         return new DescriptorHandler();
+    }
+    
+    @Override
+    public ConfigurationParser getConfigurationParser() {
+        return new ConfigurationParser();
     }
 
     @Override
