+++ com.sap.cloud.lm.sl.mta/src/test/java/com/sap/cloud/lm/sl/mta/handlers/HandlerFactoryTest.java
@@ -11,10 +11,11 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 
-import com.sap.cloud.lm.sl.mta.handlers.v1.ConfigurationParser;
-import com.sap.cloud.lm.sl.mta.handlers.v1.DescriptorHandler;
-import com.sap.cloud.lm.sl.mta.handlers.v1.DescriptorParser;
-import com.sap.cloud.lm.sl.mta.handlers.v1.DescriptorValidator;
+import com.sap.cloud.lm.sl.mta.handlers.v2.ConfigurationParser;
+import com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorHandler;
+import com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorParser;
+import com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorValidator;
+
 
 @RunWith(value = Parameterized.class)
 public class HandlerFactoryTest {
@@ -28,19 +29,15 @@ public class HandlerFactoryTest {
     public static Iterable<Object[]> getParameters() {
         return Arrays.asList(new Object[][] {
 // @formatter:off
-            // (0) Version 1.0:
-            {
-                1, null,
-            },
-            // (1) Version 2.0:
+            // (0) Version 2.0:
             {
                 2, null,
             },
-            // (2) Version 3.1:
+            // (1) Version 3.1:
             {
                 3, null,
             },
-            // (3) Unsupported version 0:
+            // (2) Unsupported version 0:
             {
                 0, "Version \"0\" is not supported",
             },
@@ -59,11 +56,8 @@ public class HandlerFactoryTest {
     public void testGetDescriptorHandler() {
         DescriptorHandler handler = new HandlerFactory(majorVersion).getDescriptorHandler();
         switch (majorVersion) {
-            case 1:
-                assertTrue(handler instanceof DescriptorHandler);
-                break;
             case 2:
-                assertTrue(handler instanceof com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorHandler);
+                assertTrue(handler instanceof DescriptorHandler);
                 break;
             case 3:
                 assertTrue(handler instanceof com.sap.cloud.lm.sl.mta.handlers.v3.DescriptorHandler);
@@ -75,11 +69,8 @@ public class HandlerFactoryTest {
     public void testGetDescriptorValidator() {
         DescriptorValidator handler = new HandlerFactory(majorVersion).getDescriptorValidator();
         switch (majorVersion) {
-            case 1:
-                assertTrue(handler instanceof DescriptorValidator);
-                break;
             case 2:
-                assertTrue(handler instanceof com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorValidator);
+                assertTrue(handler instanceof DescriptorValidator);
                 break;
             case 3:
                 assertTrue(handler instanceof com.sap.cloud.lm.sl.mta.handlers.v3.DescriptorValidator);
@@ -87,15 +78,13 @@ public class HandlerFactoryTest {
         }
     }
 
+    
     @Test
     public void testGetDescriptorParser() {
         DescriptorParser handler = new HandlerFactory(majorVersion).getDescriptorParser();
         switch (majorVersion) {
-            case 1:
-                assertTrue(handler instanceof DescriptorParser);
-                break;
             case 2:
-                assertTrue(handler instanceof com.sap.cloud.lm.sl.mta.handlers.v2.DescriptorParser);
+                assertTrue(handler instanceof DescriptorParser);
                 break;
             case 3:
                 assertTrue(handler instanceof com.sap.cloud.lm.sl.mta.handlers.v3.DescriptorParser);
@@ -107,11 +96,8 @@ public class HandlerFactoryTest {
     public void testGetConfigurationParser() {
         ConfigurationParser handler = new HandlerFactory(majorVersion).getConfigurationParser();
         switch (majorVersion) {
-            case 1:
-                assertTrue(handler instanceof ConfigurationParser);
-                break;
             case 2:
-                assertTrue(handler instanceof com.sap.cloud.lm.sl.mta.handlers.v2.ConfigurationParser);
+                assertTrue(handler instanceof ConfigurationParser);
                 break;
             case 3:
                 assertTrue(handler instanceof com.sap.cloud.lm.sl.mta.handlers.v3.ConfigurationParser);
